% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.r
\name{c_flags}
\alias{c_flags}
\title{Get C/C++ Compiler Flags for hdf5lib}
\usage{
c_flags(api = "latest")
}
\arguments{
\item{api}{A character string specifying the HDF5 API version to use.
This adds a preprocessor directive like `-DH5_USE_114_API_DEFAULT` to ensure
that the compiled code uses symbols compatible with a specific version of
the HDF5 API. This is useful for maintaining compatibility with older HDF5
versions. Supported values are `200`, `114`, `112`, `110`, `18`,and `16`. 
Defaults to `"latest"`, which corresponds to the newest supported API version.}
}
\value{
A scalar character vector containing the compiler flags (e.g., the
  `-I` path to the package's `inst/include` directory).
}
\description{
Provides the required C/C++ compiler flags to find the HDF5 header
files bundled with the `hdf5lib` package.
}
\examples{
c_flags()
c_flags(api = "114")

}
\seealso{
[ld_flags()]
}
