% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.r
\name{ld_flags}
\alias{ld_flags}
\title{Get C/C++ Linker Flags for hdf5lib}
\usage{
ld_flags(api = "latest")
}
\arguments{
\item{api}{A character string specifying the HDF5 API version. This
parameter is included for consistency with [c_flags()] and is reserved for
future use. It currently has no effect on the linker flags. Supported
values are `200`, `114`, `112`, `110`, `18`, and `16`.
Defaults to `"latest"`.}
}
\value{
A scalar character vector containing the linker flags.
}
\description{
Provides the required linker flags to link against the static HDF5
library (`libhdf5z.a`) bundled with the `hdf5lib` package.
}
\examples{
ld_flags()

}
\seealso{
[c_flags()]
}
