% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.r
\name{c_flags}
\alias{c_flags}
\title{Get C/C++ Compiler Flags for hdf5lib}
\usage{
c_flags()
}
\value{
A scalar character vector containing the compiler flags (e.g., the
  `-I` path to the package's `inst/include` directory).
}
\description{
Provides the required C/C++ compiler flags to find the HDF5 header
files bundled with the `hdf5lib` package.

This function is intended to be called from a `Makevars` file by
other R packages that link to `hdf5lib`.
}
\examples{
c_flags()
}
\seealso{
[ld_flags()]
}
