% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5lib-package.r
\docType{package}
\name{hdf5lib-package}
\alias{hdf5lib}
\alias{hdf5lib-package}
\title{hdf5lib: Headers and Static Libraries for 'HDF5'}
\description{
`hdf5lib` provides a self-contained, static build of the HDF5 C library
(version 1.14.6) and zlib (version 1.3.1).
}
\details{
This package provides no R functions and is intended for R package
developers to use in the `LinkingTo` field of their `DESCRIPTION` file.
It allows other R packages to easily link against HDF5 without
requiring users to install any system-level dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cmmr/hdf5lib}
  \item Report bugs at \url{https://github.com/cmmr/hdf5lib/issues}
}

}
\author{
\strong{Maintainer}: Daniel P. Smith \email{dansmith01@gmail.com} (\href{https://orcid.org/0000-0002-2479-2044}{ORCID})

Other contributors:
\itemize{
  \item Alkek Center for Metagenomics and Microbiome Research [copyright holder, funder]
}

}
\keyword{internal}
