% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.r
\name{ld_flags}
\alias{ld_flags}
\title{Get C/C++ Linker Flags for hdf5lib}
\usage{
ld_flags()
}
\value{
A scalar character vector containing the linker flags.
}
\description{
Provides the required linker flags to link against the static HDF5
library (`libhdf5z.a`) bundled with the `hdf5lib` package.

This function is intended to be called from a `Makevars` file by
other R packages that link to `hdf5lib`. It returns the `-L` path
to the library directory and the `-l` flags for `libhdf5` and
its system dependencies (like `pthread` and `dl`).
}
\examples{
ld_flags()
}
\seealso{
[c_flags()]
}
